only admin can manage blog

+++ apps/blog/views/delete_post.py
@@ -1,3 +1,4 @@
+from django.http import HttpResponseForbidden
 from django.shortcuts import render, redirect
 
 from ..models import Post
@@ -10,14 +11,16 @@ def delete_post(request, post_id):
     :param post_id: post id
     :return: redirect or render page
     """
-    post = Post.objects.get(id=post_id)
-    context = {"title": "Delete post", "post": post}
+    if request.user.is_staff:
+        post = Post.objects.get(id=post_id)
+        context = {"title": "Delete post", "post": post}
 
-    if request.method == "POST":
-        post.delete()
-        return redirect('blog:post_deleted')
+        if request.method == "POST":
+            post.delete()
+            return redirect('blog:post_deleted')
+        else:
+            return render(request=request,
+                          template_name="blog/delete_post.html",
+                          context={"post": context})
     else:
-        return render(request=request,
-                      template_name="blog/delete_post.html",
-                      context={"post": context})
-
+        return HttpResponseForbidden()
